Resolve import errors in lambda handlers
We needed the full import path to enable pytest on the lambda handlers,
but this broke the deployed handlers as the parent directory isn't
copied. Fixed for now by copying the parent directory to a temporary
location and uploading as src, to avoid gratuitous code in handlers
(e.g. `aggregator` doesn't need `get_and_parse_hiscores`).

This is a bit of a hack; might be nicer just to add another level of
nesting in the lambda directory. Like `lambda.aggregator` ->
`lambda.aggregator.aggregator`, etc. Will decide whether to do this
later on.

resolves #12