Add rudimentary integration tests
Going forward, it would be better to implement CI/CD and do
this in a deployment pipeline. Need to figure out the best way
to do this; it looks like all CI/CD pipelines are triggered
by a code source (which makes sense). However, we want integration
tests that can be run _prior_ to pushing the code. This must be
possible, but will require some further research. For now, a basic
integration test is available in the `run_integration_tests.py`
script.

resolves #3